Err out on `testFail
